requirements.txt
fastapi==0.68.1        # Our web framework
uvicorn==0.15.0        # ASGI server for FastAPI
requests==2.26.0       # Making API calls less painful
python-dotenv==0.19.0  # For keeping secrets secret
pytest==6.2.5          # Because we test our code!
httpx==0.28.1             #fot the http access
pydantic==1.10.13     #For data validation (used by FastAPI)
typing-extensions==4.8.0 #For improved type hinting
starlette==0.14.2  #Web framework toolkit (FastAPI dependency)
aiohttp==3.8.6  #For async HTTP requests
python-multipart==0.0.6 #For async HTTP requests
